+++ app/src/main/res/layout/fragment_blog_list.xml
