+++ app/src/main/res/layout/fragment_event_detail.xml
