+++ app/src/main/res/layout/fragment_member_detail.xml
