+++ app/src/main/res/layout/nav_header.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout
     xmlns:android="http://schemas.android.com/apk/res/android" android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/design_default_color_on_primary">
-
-</androidx.constraintlayout.widget.ConstraintLayout>
+    android:layout_height="150dp"
+    android:background="@color/purple_500"
+    />
