--- app/src/main/res/menu/nav_header_menu.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item android:id="@+id/miItem1"
-        android:title="@string/I1"/>
-
-    <item android:id="@+id/miItem2"
-        android:title="@string/I2"/>
-
-    <item android:id="@+id/miItem3"
-        android:title="@string/I3"/>
-</menu>
