--- app/src/main/res/navigation/nav_components.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/nav_components"
-    app:startDestination="@id/start_activity">
-
-    <fragment
-        android:id="@+id/start_activity"
-        android:name="com.ankurshukla.myapplication.start_activity"
-        android:label="fragment_start_activity"
-        tools:layout="@layout/fragment_start_activity" >
-        <action
-            android:id="@+id/action_start_activity_to_blog_content"
-            app:destination="@id/blog_content" />
-        <action
-            android:id="@+id/action_start_activity_to_events"
-            app:destination="@id/events" />
-        <action
-            android:id="@+id/action_start_activity_to_members"
-            app:destination="@id/members" />
-    </fragment>
-    <fragment
-        android:id="@+id/blog_content"
-        android:name="com.ankurshukla.myapplication.blog_content"
-        android:label="fragment_blog_content"
-        tools:layout="@layout/fragment_blog_content" />
-    <fragment
-        android:id="@+id/events"
-        android:name="com.ankurshukla.myapplication.events_fragment"
-        android:label="fragment_events_fragment"
-        tools:layout="@layout/fragment_events_fragment" >
-        <action
-            android:id="@+id/action_events_fragment_to_event_detail_fragment"
-            app:destination="@id/event_details" />
-    </fragment>
-    <fragment
-        android:id="@+id/event_details"
-        android:name="com.ankurshukla.myapplication.event_detail_fragment"
-        android:label="fragment_event_detail_fragment"
-        tools:layout="@layout/fragment_event_detail_fragment" />
-    <fragment
-        android:id="@+id/members"
-        android:name="com.ankurshukla.myapplication.member_fragment"
-        android:label="fragment_member_fragment"
-        tools:layout="@layout/fragment_member_fragment" >
-        <action
-            android:id="@+id/action_members_to_member_details"
-            app:destination="@id/member_details" />
-    </fragment>
-    <fragment
-        android:id="@+id/member_details"
-        android:name="com.ankurshukla.myapplication.member_detail_fragment"
-        android:label="fragment_member_detail_fragment"
-        tools:layout="@layout/fragment_member_detail_fragment" />
-</navigation>
